Report VM identity (if available) to the proxy.
This change adds a new header in the requests that the agent sends to the
proxy. That header is used to report the identity of the host VM for the
scenario that the agent is running inside of a Google Compute Engine VM.

When running outside of Google Compute Engine, this change is a no-op.

This change also cleaned up existing lint issues in the agent.