Add the ability to shim websockets directly in the agent.
This allows proxy to work with arbitrary servers even if those
servers do not themselves offer any means of replacing websockets
with HTTP calls.

The shim involves two steps:
1. When serving an HTML response, inject special Javascript code to
   replace websockets with the shim.
2. Implement a simple HTTP protocol for transferring the websocket
   requests and responses between the client and server.

It may be valuable to replace this custom shim with something more
standard like SockJS or Socket.io, but for now I am using the simplest
solution that will work.

The logic to inject the JS-side shim code is limited and defaults to
doing nothing. This means for some cases websockets will not be
shimmed by the agent and the server will still be responsible for
handling that. One particularly notable case is when an HTML response
from the server is compressed, in which case deflating the response in
order to inject the JS code would have performance implications.

The server-side logic for websocket connections uses the
gorilla/websocket library, which has been vendored in to the
repository. The version vendored in is the tag `v1.4.0`, which
is the commit 66b9c49e59c6c48f0ffce28c2d8b8a5678502c6d