Fix a subtle bug in the close logic for response forwarding
This commit fixes a logic bug in the closing of a `ResponseForwarder`
where the agent could be killed with a runtime panic if the connection
to the proxy server was interrupted.

The root cause was that the `ResponseForwarder` type tries to collect
the errors encountered by multiple goroutines and return them all as a
combined error.

It did that using a single channel from which all of the reported
errors could be obtained using a range statement. That required the
channel to be closed, and the code tried to reason about the
conditions under which it would be safe to close the channel as all
potential users had finished writing to it.

This logic overlooked one scenario; the client connection to the proxy
server getting killed. In that scenario the error channel would get
closed prior to one of the potential users being finished with it.

Rather than try to fix the reasoning, this change removes the need for
it by splitting the single error channel into three different
channels; one for each separate user.

This new structure allows us to know for certain that none of the
channels will be written to after they are closed, as we only close
them after the corresponding user has finished doing all of its work.