agent: maintain ordering of the shimmed websocket messages
Data sent via websockets are supposed to be sent in-order, but the
previous version of the client-side code for the websocket shim sent
all messages asynchronously, which could cause them to be received
out of order by the agent.

This change adds an array for storing the websocket data to be sent,
and has the client-side JS code work through this array in sequence.