Fix a bug in the locally-runnable proxy where headers were not copied
This change fixes a bug in the locally-runnable version of the
inverting proxy where the headers from a proxied response were not
copied in to the response sent back to the frontend client.

That prevents things like set-cookie from working when testing local
changes.

That, in turn, was blocking the testing of a fix to
https://github.com/google/inverting-proxy/issues/31