Add support for receiving binary WebSocket messages.
This commit extends the websocket-shim to support receiving binary
messages over the shimmed WebSocket connections.

The WebSocket-over-HTTP protocol is extended such that messages
returned by the `poll` handler may be either a string (for text
messages), or an array of a single string (for binary messages).

The shim code detects this difference between the possible types
and uses it to determine whether to emit an event with a string,
or an event with a Blob (which is constructed directly from the
array of the single string).

So far, this change only adds support for receiving binary messages,
support for sending binary messages will follow in a subsequent
commit.

This was tested by running a proxy and agent with websocket
shimming locally, pointing the agent at a locally running
Jupyter notebook server, and then running the following snippet
in a Jupyter notebook code cell:

```py
import ipywidgets

ipywidgets.Image(value='', format='png')
```

Fixes #47