agent: Don't inject a banner for downloaded HTML files
The logic in the agent to inject an HTML banner previously assumed
that any HTML response was being displayed in a browser, and thus
was suitable for banner injection as long as it was not framed.

However, that overlooks the use case of downloading an HTML file
from a server (e.g. using Jupyter Notebook's "download as HTML"
feature).

Fortunately, there is a standard response header that we can use
to distinguish this use case; the "Content-Disposition: attachment"
header. By declining to inject the banner when that response header
is present, we can still support banner injection without breaking
HTML downloads.