agent: Switch the banner logic to not use 1.14-only features.
The previous version of the banner logic used the "Header.Values"
method from the "net/http" package. That method was only added in
version 1.14 of Go, and we have some users building the agent with
version 1.12. Since we do not want to break them, this change just
changes that code to access the map contents of the header rather
than using the not-yet-available helper method.