Swap order of websocket and session handlers
This fixes an edge condition where cookies with path
restrictions are not sent when opening a websocket connection
because the path is `websocket-shim/open` when the session is restored.

This change brings the cookie handler downstream of the websocket
handler for non-shimmed requests, and restores cookies for
websocket open requests based on the correct path.