Add retries to write response calls to proxy
The goal is to avoid failing the end user request in case of a transient
network connectivity error or another internal proxy error when forwarding
a response to the proxy, assuming it succeeds on one of the next attempts.
The retries are only applied to errors that occur before the proxy has read
more than the first 4096 bytes of the response body.