Support gRPC backends.
This change fixes two categories of issues that prevented gRPC backends
from being usable from behind the inverting proxy:

1. HTTP/2 not being used for communication between the agent and backend.
2. HTTP response trailers not being propagated from the backend to the client.

The first issue only applied to the proxy agent itself, and to fix this we
add a new `--force-http2` flag to the agent that indicates that HTTP/2 should
be used even when the communication with the backend is not over TLS.

The second issue is more broadly applicable and required fixes in both the
agent and the proxy itself.

Within this change, we only made that fix for the in-memory version of the
proxy; it has not be replicated for the AppEngine version.