[agent] Refactor the `utils.ResponseForwarder` logic to be easier to understand and test.
Previously, we maintained a single `ResponseForwarder` type that was responsible for
both streaming the response body from the backend, and streaming the response in the
request body to the proxy.

Combining those two pieces of functionality into a single type made the code far too
difficult to reason about, and needlessly coupled logically separate roles together.

This change splits off a new `StreamingResponseWriter` type that is responsible for
just streaming the response back from the backend. That new type is then wrapped by
the response forwarder type so that the separate concerns (streaming *from* the backend
and streaming *to* the proxy) are fully separated and can be reasoned about in isolation.

This change also adds numerous tests for the new `StreamingResponseWriter` type.

Finally, this refactored response forwarder has slightly different behavior from the
old version with regard to the timing for when the response forwarder starts
sending out the proxy request containing the backend response. Previously, this
request was not started until we got the initial response from the backend.

That was done to eliminate the risk of the OAuth access token used in the proxy
request expiring before the response from the backend was received.

However, upon further reflection that delay is completely unnecessary; the OAuth
access token is not used during the writing of the backend response (which is
sent inside of the proxy request's body), but rather at the time that the proxy
request is initiated.

That means that it doesn't matter how long the backend takes to respond; the
request to proxy will be authenticated as soon as the request is sent, regardless
of how long of a delay there might be in writing the response body.