[agent] Fix a race condition when writing a response to a cancelled request.
When a request is cancelled prior to the response being written, we might still
get a response from the backend and try to write it.

Previously, if that happened the write of the response body would hang forever,
waiting for the response to be serialized to the no-longer extant proxy request.

With this change, that no longer happens because we catch this scenario in the
`WriteHeader` method and close the body reader... causing any subsequent writes
to immediately fail.